pyproject.toml
[project]
name = "propierty-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.15.0",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
